Add tenfold repetition to the tests with random rectangles, move test files to the separate directory
