--- TagsCloudVisualization/Program.cs
@@ -1,16 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Drawing;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace TagsCloudVisualization
-{
-    class Program
-    {
-        static void Main(string[] args)
-        {
-        }
-    }
-}
