+++ TagsCloudVisualization/TagsCloudVisualization.csproj
@@ -62,18 +62,19 @@
   </ItemGroup>
   <ItemGroup>
     <Compile Include="CircularCloudLayouter.cs" />
-    <Compile Include="CircularCloudLayouterShould.cs" />
+    <Compile Include="Tests\CircularCloudLayouterShould.cs" />
     <Compile Include="ConvexHullBuilder.cs" />
-    <Compile Include="ConvexHullBuilderShould.cs" />
-    <Compile Include="Program.cs" />
+    <Compile Include="Tests\ConvexHullBuilderShould.cs" />
+    <Compile Include="Tests\Program.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Vector.cs" />
-    <Compile Include="VectorShould.cs" />
+    <Compile Include="Tests\VectorShould.cs" />
   </ItemGroup>
   <ItemGroup>
     <None Include="App.config" />
     <None Include="packages.config" />
   </ItemGroup>
+  <ItemGroup />
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
   <Target Name="EnsureNuGetPackageBuildImports" BeforeTargets="PrepareForBuild">
     <PropertyGroup>
