+++ TagsCloudVisualization/Tests/Program.cs
@@ -1,0 +1,9 @@
+﻿namespace TagsCloudVisualization.Tests
+{
+    class Program
+    {
+        static void Main(string[] args)
+        {
+        }
+    }
+}
